refactor(functions): share safe_float_convert via api_clients
Course.ects is free text (default 'NONE', German decimal commas), so every
consumer needs the same tolerant parser. It existed only inside
apiProxy/participant_data_handler.py.

Move it to shared_libs/api_clients/numeric_utils.py and re-export it from
the package. That directory is used because the release workflow copies
exactly shared_libs/api_clients into each python bundle, so a module placed
elsewhere would work locally and fail to import in production.
