+++ functions/shared_libs/api_clients/__init__.py
@@ -3,3 +3,4 @@ from .zoom_client import ZoomClient
 from .limesurvey_client import LimeSurveyClient
 from .mattermost_client import MattermostClient
 from .storage_client import StorageClient
+from .numeric_utils import safe_float_convert
